requirements.txt
# Core dependencies
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0

# Training framework
verl-agent==0.7.0.dev0
vllm==0.11.0
xformers==0.0.32.post1

# ML frameworks
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.7.0

# PPO training
tensordict>=0.2.0
codetiming>=1.4.0

# Configuration
hydra-core>=1.3.0
omegaconf>=2.3.0
pyyaml>=6.0

# Utilities
numpy<2.0.0
pandas>=2.0.0
tqdm>=4.66.0
tensorboard>=2.14.0
wandb>=0.16.0

# Scientific computing
scipy>=1.11.0
scikit-learn>=1.3.0

# Text processing
pybind11>=2.11.0
pylatexenc>=2.10
sentencepiece>=0.1.99

# RL specific
trl>=0.7.0
